Move persist store to stores. Move websocket to utilities
