+++ src/lib/stores/data-converter-config.ts
@@ -1,5 +1,5 @@
 import { writable } from 'svelte/store';
-import { persistStore } from '$lib/utilities/persist-store';
+import { persistStore } from '$lib/stores/persist-store';
 
 export const dataConverterPort = persistStore('port', null);
